Dispatch utility_functions.php. Closed #141
